moved model struct to ports
