+++ services/carbon-intensity-provider/adapters/notifier/notifier.go
@@ -4,7 +4,7 @@ import (
 	"context"
 	"log"
 
-	"github.com/informatik-mannheim/cmg-ss2025/services/carbon-intensity-provider/model"
+	"github.com/informatik-mannheim/cmg-ss2025/services/carbon-intensity-provider/ports"
 )
 
 type Notifier struct{}
@@ -13,6 +13,6 @@ func New() *Notifier {
 	return &Notifier{}
 }
 
-func (d *Notifier) CarbonIntensityProviderChanged(data model.CarbonIntensityData, ctx context.Context) {
+func (d *Notifier) CarbonIntensityProviderChanged(data ports.CarbonIntensityData, ctx context.Context) {
 	log.Printf("[Notification] Notifier triggered for zone: %s, intensity: %.2f", data.Zone, data.CarbonIntensity)
 }
