+++ services/carbon-intensity-provider/adapters/repo-in-memory/repo.go
@@ -3,12 +3,11 @@ package repo_in_memory
 import (
 	"context"
 
-	"github.com/informatik-mannheim/cmg-ss2025/services/carbon-intensity-provider/model"
 	"github.com/informatik-mannheim/cmg-ss2025/services/carbon-intensity-provider/ports"
 )
 
 type Repo struct {
-	carbonIntensityProviders map[string]model.CarbonIntensityData
+	carbonIntensityProviders map[string]ports.CarbonIntensityData
 }
 
 // Ensure Repo implements ports.Repo
@@ -16,25 +15,25 @@ var _ ports.Repo = (*Repo)(nil)
 
 func NewRepo() *Repo {
 	return &Repo{
-		carbonIntensityProviders: make(map[string]model.CarbonIntensityData),
+		carbonIntensityProviders: make(map[string]ports.CarbonIntensityData),
 	}
 }
 
-func (r *Repo) Store(data model.CarbonIntensityData, ctx context.Context) error {
+func (r *Repo) Store(data ports.CarbonIntensityData, ctx context.Context) error {
 	r.carbonIntensityProviders[data.Zone] = data
 	return nil
 }
 
-func (r *Repo) FindById(id string, ctx context.Context) (model.CarbonIntensityData, error) {
+func (r *Repo) FindById(id string, ctx context.Context) (ports.CarbonIntensityData, error) {
 	data, ok := r.carbonIntensityProviders[id]
 	if !ok {
-		return model.CarbonIntensityData{}, ports.ErrCarbonIntensityProviderNotFound
+		return ports.CarbonIntensityData{}, ports.ErrCarbonIntensityProviderNotFound
 	}
 	return data, nil
 }
 
-func (r *Repo) FindAll(ctx context.Context) ([]model.CarbonIntensityData, error) {
-	result := make([]model.CarbonIntensityData, 0, len(r.carbonIntensityProviders))
+func (r *Repo) FindAll(ctx context.Context) ([]ports.CarbonIntensityData, error) {
+	result := make([]ports.CarbonIntensityData, 0, len(r.carbonIntensityProviders))
 	for _, data := range r.carbonIntensityProviders {
 		result = append(result, data)
 	}
