+++ services/carbon-intensity-provider/core/service.go
@@ -3,7 +3,6 @@ package core
 import (
 	"context"
 
-	"github.com/informatik-mannheim/cmg-ss2025/services/carbon-intensity-provider/model"
 	"github.com/informatik-mannheim/cmg-ss2025/services/carbon-intensity-provider/ports"
 )
 
@@ -19,16 +18,16 @@ func NewCarbonIntensityService(repo ports.Repo, notifier ports.Notifier) *Carbon
 	}
 }
 
-func (s *CarbonIntensityService) GetCarbonIntensityByZone(zone string, ctx context.Context) (model.CarbonIntensityData, error) {
+func (s *CarbonIntensityService) GetCarbonIntensityByZone(zone string, ctx context.Context) (ports.CarbonIntensityData, error) {
 	return s.repo.FindById(zone, ctx)
 }
 
-func (s *CarbonIntensityService) GetAvailableZones(ctx context.Context) []model.Zone {
+func (s *CarbonIntensityService) GetAvailableZones(ctx context.Context) []ports.Zone {
 	data, _ := s.repo.FindAll(ctx)
 
-	zones := make([]model.Zone, 0, len(data))
+	zones := make([]ports.Zone, 0, len(data))
 	for _, item := range data {
-		zones = append(zones, model.Zone{
+		zones = append(zones, ports.Zone{
 			Code: item.Zone,
 			Name: item.Zone,
 		})
@@ -37,7 +36,7 @@ func (s *CarbonIntensityService) GetAvailableZones(ctx context.Context) []model.
 }
 
 func (s *CarbonIntensityService) AddOrUpdateZone(zone string, intensity float64, ctx context.Context) error {
-	provider := model.CarbonIntensityData{
+	provider := ports.CarbonIntensityData{
 		Zone:            zone,
 		CarbonIntensity: intensity,
 	}
