+++ services/carbon-intensity-provider/ports/notifier.go
@@ -2,10 +2,8 @@ package ports
 
 import (
 	"context"
-
-	"github.com/informatik-mannheim/cmg-ss2025/services/carbon-intensity-provider/model"
 )
 
 type Notifier interface {
-	CarbonIntensityProviderChanged(data model.CarbonIntensityData, ctx context.Context)
+	CarbonIntensityProviderChanged(data CarbonIntensityData, ctx context.Context)
 }
