+++ services/carbon-intensity-provider/ports/service.go
@@ -2,12 +2,10 @@ package ports
 
 import (
 	"context"
-
-	"github.com/informatik-mannheim/cmg-ss2025/services/carbon-intensity-provider/model"
 )
 
 // CarbonIntensityProvider defines the service interface for managing carbon intensity data.
 type CarbonIntensityProvider interface {
-	GetCarbonIntensityByZone(zone string, ctx context.Context) (model.CarbonIntensityData, error)
-	GetAvailableZones(ctx context.Context) []model.Zone
+	GetCarbonIntensityByZone(zone string, ctx context.Context) (CarbonIntensityData, error)
+	GetAvailableZones(ctx context.Context) []Zone
 }
